upgrade DFS pathfinding algo
